// src/app/Modules/layout-settings/Model/ruleengine.model.ts
export class RuleEngine {
    public ruleId: string = '';
    public layoutCategoryId: string = '';
    public roomName: string = '';
    public roomNumber: string = '';
    public layoutPositionId: number;
    responseModel: RuleEngine[];
    keyword: string;
    activeStatus: string;
  statusMessage: string;
  pageSize?: number;
  page?: number;
  count?:number
  }
  